use proper pgpass location on Windows
